+++ examples/serial_vcom.rs
@@ -12,19 +12,18 @@ extern crate cortex_m_rt as rt;
 extern crate nb;
 extern crate panic_semihosting;
 
-extern crate stm32l432xx_hal as hal;
+extern crate stm32l4_hal as hal;
 // #[macro_use(block)]
 // extern crate nb;
 
 use cortex_m::asm;
 use crate::hal::prelude::*;
 use crate::hal::serial::Serial;
-use crate::hal::stm32l4::stm32l4x2;
 use crate::rt::ExceptionFrame;
 
 #[entry]
 fn main() -> ! {
-    let p = stm32l4x2::Peripherals::take().unwrap();
+    let p = hal::stm32::Peripherals::take().unwrap();
 
     let mut flash = p.FLASH.constrain();
     let mut rcc = p.RCC.constrain();
@@ -44,7 +43,7 @@ fn main() -> ! {
 
      // TRY using a different USART peripheral here
     let serial = Serial::usart2(p.USART2, (tx, rx), 9_600.bps(), clocks, &mut rcc.apb1r1);
-    let (mut tx, mut rx) = serial.split();
+    let (mut tx, _) = serial.split();
 
     let sent = b'X';
 
@@ -58,7 +57,7 @@ fn main() -> ! {
     block!(tx.write(sent)).ok();
     block!(tx.write(sent)).ok();
     block!(tx.write(sent)).ok();
-    
+
     // when using virtual com port for recieve, causes a framing error
     // let received = block!(rx.read()).unwrap();